+++ api/src/main/java/com/example/coupangclone/security/userdetails/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.coupangclone.userdetails;
+package com.example.coupangclone.security.userdetails;
 
 import com.example.coupangclone.entity.user.User;
 import com.example.coupangclone.repository.user.UserRepository;
